Add copy of /atom.xml as /rss.
The reason for this is that at this blog's previous home on Tumblr,
the feed was /rss, and it has existing subscribers. I could put a dummy
RSS feed at /rss telling people to update their feeds, but if redirecting
them to the current atom.xml works, that's less intrusive. Meanwhile I
don't know how to literally redirect them -- using Github Pages hosting
I don't control redirects, and if I leave rss as a symlink to atom.xml
Octopress generate phase ignores it -- but just copying the atom.xml
template to rss seems to work ok. So that's what's done here.